[REFACTOR]: replace context with useSupabase hook
